requirements.txt
fastapi
uvicorn
line-bot-sdk
trafilatura
google-generativeai
python-dotenv
youtube-transcript-api